:bento: Chore: best, worst api data class 정의

+++ app/src/main/java/com/example/snack4diet/api/bestWorst/Data.kt
@@ -2,10 +2,10 @@ package com.example.snack4diet.api.bestWorst
 
 data class Data(
     val best: List<Best>,
-    val calorieScore: Int,
-    val carbohydrateScore: Int,
-    val fatScore: Int,
-    val proteinScore: Int,
-    val totalScore: Int,
+    val calorieScore: Double,
+    val carbohydrateScore: Double,
+    val fatScore: Double,
+    val proteinScore: Double,
+    val totalScore: Double,
     val worst: List<Worst>
 )
